backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
openai
python-dotenv
pydantic
google-generativeai
passlib[bcrypt]
email-validator